Refactor updaters - MLN and CG now have a common implementation
