+++ deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/updater/UpdaterCreator.java
@@ -4,7 +4,9 @@ import org.deeplearning4j.nn.api.Model;
 import org.deeplearning4j.nn.conf.NeuralNetConfiguration;
 import org.deeplearning4j.nn.conf.Updater;
 import org.deeplearning4j.nn.conf.layers.Layer;
+import org.deeplearning4j.nn.graph.ComputationGraph;
 import org.deeplearning4j.nn.multilayer.MultiLayerNetwork;
+import org.deeplearning4j.nn.updater.graph.ComputationGraphUpdater;
 import org.nd4j.linalg.learning.GradientUpdater;
 
 /**
@@ -19,6 +21,8 @@ public class UpdaterCreator {
     public static org.deeplearning4j.nn.api.Updater getUpdater(Model layer) {
         if (layer instanceof MultiLayerNetwork) {
             return new MultiLayerUpdater((MultiLayerNetwork) layer);
+        } else if(layer instanceof ComputationGraph){
+            return new ComputationGraphUpdater((ComputationGraph)layer);
         } else {
             return new LayerUpdater();
         }
